Clean up root directory organization
- Move check_app_health.py to scripts/ (updated version)
- Move test_image.jpg to tests/ directory
- Update requirements.txt
