+++ requirements.txt
@@ -1,7 +1,7 @@
+aiofiles==24.1.0
 aiohappyeyeballs==2.6.1
 aiohttp==3.12.14
 aiosignal==1.4.0
-aiofiles==24.1.0
 alembic==1.16.4
 annotated-types==0.7.0
 anthropic==0.60.0
@@ -53,6 +53,8 @@ fastapi==0.109.2
 fastavro==1.11.1
 filelock==3.18.0
 flatbuffers==25.2.10
+flexcache==0.3
+flexparser==0.4
 frozenlist==1.7.0
 fsspec==2025.7.0
 fuzzywuzzy==0.18.0
@@ -85,6 +87,7 @@ hypothesis==6.136.3
 idna==3.10
 importlib_metadata==8.7.0
 importlib_resources==6.5.2
+ingredient_parser_nlp==2.1.1
 iniconfig==2.1.0
 instructor==1.10.0
 ipykernel==6.30.0
@@ -94,6 +97,7 @@ itsdangerous==2.2.0
 jedi==0.19.2
 Jinja2==3.1.6
 jiter==0.10.0
+joblib==1.5.1
 json5==0.12.0
 json_repair==0.25.2
 jsonpatch==1.33
@@ -134,6 +138,7 @@ multidict==6.6.3
 mypy_extensions==1.1.0
 nest-asyncio==1.6.0
 networkx==3.5
+nltk==3.9.1
 nodeenv==1.9.1
 numpy==1.26.4
 oauthlib==3.3.1
@@ -163,6 +168,7 @@ pdfminer.six==20250506
 pdfplumber==0.11.7
 pexpect==4.9.0
 pillow==11.3.0
+Pint==0.24.4
 pip-api==0.0.34
 platformdirs==4.3.8
 playwright==1.44.0
@@ -200,6 +206,7 @@ pytest==8.4.1
 pytest-asyncio==1.1.0
 pytest-mock==3.14.1
 pytest-vcr==1.0.2
+python-crfsuite==0.9.11
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
 python-jose==3.3.0
@@ -233,6 +240,7 @@ sounddevice==0.5.1
 soundfile==0.13.1
 soupsieve==2.7
 SQLAlchemy==2.0.41
+sqlmodel==0.0.24
 sse-starlette==3.0.2
 stack-data==0.6.3
 stagehand==0.5.0
